+++ config/filament-latex.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'navigation-label' => 'Filament Latex',
+    'navigation-icon' => 'vendor.filament-latex.latex',
+    'navigation-group' => null,
+    'list-page-title' => 'Latex Documents',
+    'create-button-label' => 'Create Document',
+    'create-page-title' => 'Create Latex Document',
+    'user-model' => 'App\Models\User',
+];